Fix python installation, formatting, readme
